finished flatter ex01_04 - just program class
