+++ Ex01 Guy 322372681 Roy 322718388/Ex01_04/Ex01_04.csproj
@@ -36,8 +36,6 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConsoleHandler.cs" />
-    <Compile Include="StringCheckingFunctions.cs" />
     <Compile Include="Program.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
